Use `_.merge` instead of `Object.assign` for merging user config with generated config object
`Object.assign` will, in most cases, end up completely overriding the generated config object with the user one, which means that most of the time we'll end up with a "corrupt" config object that won't work.

Using `_.merge` instead allows the user to define a "partial" config. There may be better ways of handling this, like passing along the generated config to the user, so they can do what they want, but using `_.merge` is by far the simplest solution for now.